+++ packages/uring_api/src/_uring_api/uring_api_capi_impl.h
@@ -69,6 +69,7 @@ int UringApiCapi_RingResetServing(PyObject *ring);
 int UringApiCapi_RingSubmit(PyObject *ring, int *submitted);
 int UringApiCapi_CompletionCheck(PyObject *completion);
 PyObject *UringApiCapi_CompletionUserData(PyObject *completion);
+int UringApiCapi_CompletionSetUserData(PyObject *completion, PyObject *value);
 int UringApiCapi_CompletionRes(PyObject *completion, int *value);
 int UringApiCapi_CompletionFlags(PyObject *completion, unsigned int *value);
 int UringApiCapi_CompletionSequence(PyObject *completion, unsigned long long *value);